requirements.txt
torch
torch_geometric
numpy
pandas
scikit-learn
scipy
pyyaml
tqdm
matplotlib
pytest
